Addition of a section on potential pitfalls
I'm not sure this is best done as a separate section, but I wanted a place to compile notes along these lines, for future discussion. 

From my very preliminary perusal, the existing draft seems to focus on principles more than practices. If so (and I do need to read it more carefully), my thought would be to add sections that are more like recipes that we find effective.  As noted elsewhere, we may view these as standard operating procedures (SOPs) more, perhaps, than best practices, as multiple practices might be in a tie for "best".